Implement instancing directives, various other fixes
Implement parse_ObjectBegin(), parse_ObjectEnd() and parse_ObjectInstance().

Refer to other scene items using uint32_t indexes instead of pointers. This saves a lot of memory on large scenes like the Moana island. It will also make it easier to implement deferred name lookups for textures, materials, etc: before we would have had to allocate a new object, then find all places where we were pointing at the old object and update them; now we just have to change the value at the relevant index.

Rename the kInvalidTexture constant to kInvalidIndex, so we can use it in more places.

Fix a bug in parse_TransformBegin() where it was pushing the attribute stack instead of the transform stack.

Save the current transform when creating a camera too (this was omitted before, which was a bug).

Print counts of each different shape type.