Deferred name resolution for textures and materials, plus various other fixes.
We're now successfully parsing every PBRT file I can find to test with, including: all of pbrt-v3-scenes, all of Benedikt Bitterli's scenes and Disney's Moana island scene.

Implement deferred name resolution for materials, textures and mediums. Print an error message if we reach the WorldEnd statement and there are any textures or mediums which are still undefined. For materials, any undefined references get replaced with a default MatteMaterial.

Bug fix: creating a camera is supposed to also create a coordinate system called "camera".

Bug fix: incorrect parsing of ActiveTransform args. The args here are keywords, not strings (i.e. not surrounded by double quotes). Extended `parse_args()` to accept a 'k', meaning the arg is a keyword-style enum. Added `which_keyword` to the tokenizer to support this.

Bug fix: an empty string is a valid alias for the "none" material.

Bug fix: off-by-one error for texture types. This was because there are two different types of checkerboard texture.

Bug fix: the count of values in a spectrum should have been the number of floats, but we were returning the number of *pairs* of floats.

Added a snippet of the buffer contents to the Error object.

Added optional default values to use as a fallback when an enum value isn't recognised. This was needed by the Material statement, which is supposed to fall back to the "matte" material if the material type isn't recognised. Some of the test scenes fail to parse without this because they seem to be using the material name where the material class is expected.