WIP on loading PLY files
All shapes now have a `triangle_mesh()` method which will triangulate the shape and return either a new TriangleMesh instance to represent it, or null if we can't triangulate the shape. I've added a work-in-progress implementation for PLYMesh shapes which attempts to load the PLY file and extract the triangle mesh data from it.

The scene now has several new helper methods for end users to call:
- `to_triangle_mesh` replaces the shape at a particular index in the list with its triangulation
- `shapes_to_triangle_mesh` calls to_triangle_mesh for all shapes whose type belongs to a user-provided set of types.
- `load_all_ply_meshes` calls to_triangle_mesh for all shapes with type == plymesh.

Note that we do *not* automatically load ply meshes. This is deliberate: it gives users the freedom to ignore ply meshes, or to load them in parallel using their own threading system, if they choose. The helper methods above are provided so that the simple single-threaded case is no more than one extra line of code.

The PLY file format is fully supported including:
- ascii, binary little-endian and binary big-endian file types;
- comments (in the header, and in the data section as well for ascii files);

TriangleMeshes are not being populated with data from a PLY file yet (the data is being loaded, but we're not using it). That is still work-in-progress.