Extract per-vertex data from PLY files, plus various fixes
Extract vertex data from PLY mesh files and store it in the triangle mesh. Added functions for extracting data from the current element of the PLYReader. Using this to get the vertex position & optionally vertex normal & tex coords from the PLY file.

Resolve PLY mesh filenames relative to the original file, not the current file. Added a method to the Tokenizer for getting the original filename, to enable this. This fixes a bug when you have an included file in a different directory with plymesh definitions.

Treat more characters as "safe" chars to end the read buffer on. Anything <= 32 or  >= 127 is fair game now. This includes everything we allowed before, as well as all ASCII control chars and all extended ASCII chars. This helps fix a performance issue when parsing PLY files: previously, when the end of the read buffer was in the binary data section of the file, we would sometimes have to backtrack a long way looking for a safe char. This change means that search will generally terminate much sooner.

We now treat '\n' as unsafe for ending a PLY file buffer. This fixes a bug where the `next_line()` method would move the current buffer position to the end of the buffer without calling `refill_buffer()`, if the last character in the buffer was a newline.

Currently all of pbrt-v3-scenes parses successfully except for simple/dof-dragons.pbrt, which fails while loading simple/geometry/dragon.ply (which is in ASCII format). All of Benedikt Bitterli's scenes parse correctly, as does the Moana island.

I'm not extracting face indices from PLY meshes yet.