Extract face data for triangles and quads from PLY files
We now populate triangle mesh indices when loading a PLY file. Faces with 4 verts are turned into a pair of triangles.

We do not currently support faces with more than 4 verts. This doesn't seem to be a problem for any of the scenes from pbrt-v3-scenes or Benedikt Bitterli's collection, but it could occur in theory so the loader is not technically complete or correct until that's handled.

Also updated the README and TODO files.