Fix name scoping & override behaviour for textures and materials
Texture and material names are only supposed to be in scope from their point of definition through to the end of their enclosing AttributeBegin or ObjectBegin block. I've changed the lookup behaviour so that:
1. each entry in the attributes stack stores the list of texture/material indices that were defined in it; and
2. we *only* search through the items referenced in the attribute stack.

Float textures and spectrum textures are supposed to exist in a different namespace. Each entry in the Attribute stack stores separate lists of float and spectrum textures now & I modified the `find_texture` method to take a parameter saying which list to search in.

Removed the name resolver classes. We no longer do any deferred name resolution, lookup is always done immediately at the point of reference. Modified some methods to allow name lookup to fail without making the whole parse fail, to match PBRT's behaviour more closely.

Modified `find_object` and find_medium` to search through their respective arrays in reverse. This means we'll find the most recently defined item with a particular name first.

Modified `parse_ObjectInstance` to simply do nothing if the object name isn't defined. This matches PBRT's behaviour. It also now disallows ObjectInstance directives inside of an ObjectBegin/End block, to match PBRT's behaviour.

In addition to starting a new instance ObjectBegin is also supposed to push a new attribute scope, just like AttributeBegin. Similarly ObjectEnd pops an attribute scope. just like AttributeEnd. I modified the parse_ObjectBegin and parse_ObjectEnd methods to have this effect.

Removed the snapshot of buffer contents from the error message, it was only really useful for debugging & not as useful as inspecting the buffer in the debugger.