Store materials per shape not per instance & remove nested instancing support
Materials are supposed to be settable per shape, not per instance. Previously we had this backwards. We now have a material index in the Shape struct and have removed it from the Instance struct.

Shapes were previously storing the index of the instance they were part of, if any. This is unnecessary information & can waste a fair bit of memory on scenes with a lot of shapes (e.g. the Moana island) so I've removed it.

Removed support for nested instancing. This isn't actually supported by PBRT itself, so any file that has an ObjectInstance directive inside an ObjectBegin/End block is in error.