Make the has_material_override check more data-driven
Trying to reduce the number of calls to find_param, which iterates over the entire param set each time (stopping early if it finds something). We now iterate once over the list of parsed params, checking each as we go. If it's a float param, we look for the param name in a list of known float params for the target material type.