Don't ever rewind the file pointer while parsing text
Modified the logic in refill_buffer for both Tokenizer and PLYReader so that we:
(a) don't need to use fseek after trimming the read buffer back to the first safe char; and
(b) don't need to use ftell to keep track of the current file offset for the read buffer.

This doesn't make much difference when reading from an SSD (or when the data is already in the OS's disk cache), but can help when reading from a spinning disk.