Fix bugs in transforms, float parsing, .ply loading; add an "instanced" flag to shapes
The instanced flag indicates whether the shape is part of an object or not. When the flag is true the shape should not be added directly to the scene: it should only ever be added as part of instantiating it's containing object. When the flag is false, the shape should be added directly to the scene.

Our parsing was not handling negative floats correctly, it was ignoring the - sign and returning a positive number. Oops!

Fixed some bugs in `Mat4::rotate`: two incorrect indexes into the rotation axis array, one sign error.

Fixed a bug in the `dir` calculation in `Mat4::lookAt`: incorrect index when calculating the z component.

Made some changes to the matrix calculation in `Mat4::lookAt`, but it's still not working properly yet.

Found out that PBRT's Transform and ConcatTransform directives take their arguments in column-major order, despite PBRT's matrix class using row major order. Strange. Anyway, I adjusted the relevant Mat4 methods to expect the values in column-major order.

Fixed incorrect number of bytes calculation when copying indices from the PLY mesh face element into a TriangleMesh's indices member. I was only copying 1/3rd as many indices as I should have been (i.e. equal to the number of triangles, not the number of indices).

Fixed a bug where the top-most pbrt file wasn't being closed after parsing.

Fixed a bug in parse_Translate where we were not initialising the `args` array that was passed in to `m_transforms->translate`. Oops.