Add a matrix inverse method & use it fix up camera transform handling
Added a method to compute the inverse of a 4x4 matrix, using the cofactor method & pulling out common subexpressions. The need for this is because cameras in PBRT store their camera-to-world transformation, but the file format specifies that the current transformation matrix is used as the world-to-camera transform for Camera directives. I'd overlooked this earlier and it was messing up my camera positioning! I added a NOTES.md doc whose initial contents are just my working out for the inverse calculation.

Note that this feels like a wart in PBRT itself. It makes positioning cameras awkward because you're effectively transforming the world rather than the camera, so every transform has to be the inverse of what you would expect. For example, to move the camera to P you have to translate by -P.  The current transform should simply be used to specify the camera-to-world transform directly. This would have the added advantages of (a) being consistent with everything else in PBRT; and (b) making the inverse calculations unnecessary.