Hide the Parser and Tokenizer classes from end users, add a Loader class to wrap them.
These two classes contain a lot of detail that is irrelevant to end users, so it's better to have them hidden. I added the Loader class, which acts as a very thin wrapper around the Parser, as the thing end users should work with. This enabled me to move the declarations for the Tokenizer, Parser and several associated constants and helper types into the .cpp file.

I've updated the example code in the README and in the header file to use the new Loader class. I've also updated the command line app to use it.