Fix potential seg fault in the Tokenizer destructor
If the Tokenizer is destroyed without `open_file()` having been called on it, the m_fileData member will be null so there will be a seg fault when we attempt to iterate over it to free the individual members. The fix is just to check whether it's non-null before doing that.