Remove hardcoded length limits for some strings
Added the `token()` method to the tokenizer, which allows direct access to the contents of the Tokenizer's read buffer, starting from the current token. This along with the new `push_string` helper method on the Parser class, means we can copy a parsed string directly into the temp data buffer without having to copy it into a fixed-size buffer first as an intermediate step.

This change removes hardcoded limits for string literals, but we do still have limits on other things for now:
- identifiers are limited to 512 chars
- path names are limited to 512 chars before resolution, 1024 chars after.