Fix a memory leak for parameter names
Every time we encounter a parameter in the pbrt file we keep a copy of its name, but the name was never being deleted. Depending on the scene, this could have been wasting a sizeable amount of memory: tens or even hundreds of MB.

The fix here is to intern the name strings, so that we only ever have a single copy of each unique parameter name which we reuse when parsing a new set of preoprties, instead of allocating them separately. We clean up the interned strings in the Parser's destructor.