Don't use small stack-allocated strings for filename processing
We now use a preallocated temporary buffer for filename processing instead of small local stack-allocated buffers. The temporary buffer is 64 KB long, which is double the maximum path length of any current operating system.

This fixes #2

The same buffer is used by the tokenizer and the parser, so we have to be a bit careful about how we use it. In particular, I had to add a check to resolve_file for whether the `filename` param points to an address in the output `buf` param.